+++ apps/manifest-demo/webpack-host/webpack.config.js
@@ -1,6 +1,6 @@
 const path = require('path');
-const { registerPluginTSTranspiler } = require('nx/src/utils/nx-plugin.js');
-registerPluginTSTranspiler();
+// const { registerPluginTSTranspiler } = require('nx/src/utils/nx-plugin.js');
+// registerPluginTSTranspiler();
 const {
   ModuleFederationPlugin,
 } = require('@module-federation/enhanced/webpack');
@@ -29,10 +29,22 @@ module.exports = composePlugins(withNx(), withReact(), (config, context) => {
       shared: {
         lodash: {},
         antd: {},
-        react: {},
-        'react/': {},
-        'react-dom': {},
-        'react-dom/': {},
+        'react/': {
+          singleton: true,
+          requiredVersion: '^18.3.1',
+        },
+        react: {
+          singleton: true,
+          requiredVersion: '^18.3.1',
+        },
+        'react-dom': {
+          singleton: true,
+          requiredVersion: '^18.3.1',
+        },
+        'react-dom/': {
+          singleton: true,
+          requiredVersion: '^18.3.1',
+        },
       },
       experiments: { federationRuntime: 'hoisted' },
       runtimePlugins: [path.join(__dirname, './runtimePlugin.ts')],
@@ -62,6 +74,7 @@ module.exports = composePlugins(withNx(), withReact(), (config, context) => {
     minimize: false,
     moduleIds: 'named',
     chunkIds: 'named',
+    splitChunks: false,
   };
   config.output.publicPath = 'http://localhost:3013/';
   return config;