+++ packages/sdk/src/logger.ts
@@ -1,8 +1,15 @@
 import { BROWSER_LOG_KEY, BROWSER_LOG_VALUE } from './constant';
 import { isBrowserEnv, isDebugMode } from './env';
-import { safeToString } from './utils';
 
 const DEBUG_LOG = '[ FEDERATION DEBUG ]';
+
+function safeToString(info: any): string {
+  try {
+    return JSON.stringify(info, null, 2);
+  } catch (e) {
+    return '';
+  }
+}
 
 function safeGetLocalStorageItem() {
   try {