+++ packages/retry-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/retry-plugin
 
+## 0.6.9
+
+## 0.6.8
+
+## 0.6.7
+
 ## 0.6.6
 
 ## 0.6.5